Updates to Navbar CSS, Login/Register Pages, Global Colors and App.js
Wrapped Routes inside an AppLayout component and used useLocation() function to find current path, this way it can condition to hide the Navbar when the pages switch to Login or Register Page.

Deleted blue2 image used for login/register page.

Removed </Navbar> from login and register pages so as not to clash with the App.js logic.

Worked on the look for Login/Register Pages

Changed global color brand-gray to a better shade of gray